Add Food Selection save nutrients dialog use case

+++ app/src/main/java/gr/dipae/thesisfitnessapp/domain/diet/entity/Food.kt
@@ -8,7 +8,7 @@ data class Food(
     val proteins: Double,
     val carbohydrates: Double,
     val fats: Double,
-    val calories: Double,
+    val calories: Long,
     val water: Double,
     val fiber: Double
 )
